Refactor: Restructure client components and sidebar layout
- 클라이언트 컴포넌트의 경우, 부모 컴포넌트에 'use client'가 명시되어있어도 'use client'하는 것으로 수정
- 사이드바 레이아웃 수정(패딩 및 마진 변경, 축소 모드일 때 그룹 라벨 숨기기)
