--- apps/web/src/app/(dashboard)/_clientBoundary/DashboardHeaderClient/index.tsx
@@ -1,42 +1,0 @@
-'use client';
-
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbList,
-  BreadcrumbPage,
-  BreadcrumbSeparator,
-} from '@workspace/ui/components/breadcrumb';
-import { Separator } from '@workspace/ui/components/separator';
-import { SidebarTrigger } from '@workspace/ui/components/sidebar';
-import { usePathname } from 'next/navigation';
-
-import { getPathData } from '../../_helpers/utils/dashboardHeaderUtils';
-
-export function DashboardHeaderClient() {
-  const pathname = usePathname();
-  const headerData = getPathData(pathname);
-
-  return (
-    <header className="group-has-data-[collapsible=icon]/sidebar-wrapper:h-12 flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear">
-      <div className="flex items-center gap-2 px-4">
-        <SidebarTrigger className="-ml-1" />
-        <Separator
-          orientation="vertical"
-          className="mr-2 data-[orientation=vertical]:h-4"
-        />
-        <Breadcrumb>
-          <BreadcrumbList>
-            <BreadcrumbItem className="block">
-              {headerData.category}
-            </BreadcrumbItem>
-            <BreadcrumbSeparator className="block" />
-            <BreadcrumbItem>
-              <BreadcrumbPage>{headerData.title}</BreadcrumbPage>
-            </BreadcrumbItem>
-          </BreadcrumbList>
-        </Breadcrumb>
-      </div>
-    </header>
-  );
-}
