+++ apps/web/src/app/(dashboard)/_clientBoundary/UserAccountClient/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Avatar, AvatarFallback } from '@workspace/ui/components/avatar';
 import {
   DropdownMenu,
@@ -18,7 +20,7 @@ import { BadgeCheck, CreditCard, LogOut } from 'lucide-react';
 
 import { type UserAccountInfo } from '../../_helpers/types';
 
-export const UserAccount = ({ user }: { user: UserAccountInfo }) => {
+export const UserAccountClient = ({ user }: { user: UserAccountInfo }) => {
   const { isMobile } = useSidebar();
 
   return (
@@ -31,7 +33,7 @@ export const UserAccount = ({ user }: { user: UserAccountInfo }) => {
               className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground">
               <Avatar className="h-8 w-8 rounded-lg">
                 <AvatarFallback className="rounded-lg">
-                  {user.name} {/* TODO: 아바타로 변경해야 함 */}
+                  {user.name[0]} {/* TODO: 아바타로 변경해야 함 */}
                 </AvatarFallback>
               </Avatar>
               <div className="grid flex-1 text-left text-sm leading-tight">
@@ -49,7 +51,7 @@ export const UserAccount = ({ user }: { user: UserAccountInfo }) => {
               <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
                 <Avatar className="h-8 w-8 rounded-lg">
                   <AvatarFallback className="rounded-lg">
-                    {user.name} {/* TODO: 아바타로 변경해야 함 */}
+                    {user.name[0]} {/* TODO: 아바타로 변경해야 함 */}
                   </AvatarFallback>
                 </Avatar>
                 <div className="grid flex-1 text-left text-sm leading-tight">
