+++ apps/web/src/app/(dashboard)/_components/DashboardHeader/index.tsx
@@ -1,0 +1,19 @@
+import { Separator } from '@workspace/ui/components/separator';
+import { SidebarTrigger } from '@workspace/ui/components/sidebar';
+
+import { HeaderBreadcrumbClient } from '../../_clientBoundary/HeaderBreadcrumbClient';
+
+export const DashboardHeader = () => {
+  return (
+    <header className="group-has-data-[collapsible=icon]/sidebar-wrapper:h-12 flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear">
+      <div className="flex items-center gap-2 px-4">
+        <SidebarTrigger className="-ml-1" />
+        <Separator
+          orientation="vertical"
+          className="mr-2 data-[orientation=vertical]:h-4"
+        />
+        <HeaderBreadcrumbClient />
+      </div>
+    </header>
+  );
+};
